requirements.txt
numpy
pandas
opencv-python
pdf2image
openai
pinecone
python-dotenv
streamlit